Node commit
Note to self:
- This really needs to be node app where when you move the electrode
around (to different RFs) it returns all the relevant stimulus maps
loaded off the server. This off-loads all the calculations to be
pre-calculated and all the memory is runtime. Instead of having to do
complicated dot product multiplication in real time this would be a lot
smarter... Something to think about when implementing the full
simulation models in the future for sure.